pyproject.toml
[tool.poetry]
name = "qubership-pipelines-common-library"
version = "2.0.0"
description = "Qubership Pipelines common library"
authors = ["Qubership"]
readme = "README.md"
license = "Apache-2.0"
packages = [{include = "qubership_pipelines_common_library"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
click = "^8.1.7"
requests = "^2.32.3"
urllib3 = "^2.2.3"
python-jenkins = "^1.8.2"
GitPython = "^3.1.43"
http-exceptions = "^0.2.10"
python-gitlab = "^4.13.0"
minio = "^7.2.12"
kubernetes = "^34.1.0"
webexpythonsdk = "2.0.1"
ghapi = "^1.0.6"
boto3 = "^1.39.4"
google-cloud-artifact-registry = "^1.16.1"
rich = "^14.2.0"
# jira = "3.10.5"

[tool.poetry.group.test.dependencies]
pytest = "^6.0.0"

[tool.pytest.ini_options]
markers = [
    "integration: tests requiring external services (deselect with '-m \"not integration\"')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
